+++ EveESIClient/Contacts/ContactsClient.cs
@@ -37,7 +37,7 @@ namespace EveESIClient.Contacts
         /// <param name="characterId">An EVE character ID</param>
         /// <param name="contactIds">A list of contacts to delete</param>
         /// <returns>IsSuccessful will be true on successfull call</returns>
-        public ESIResponse<object> DeleteCharacterContacts(string authToken, Int64 characterId, Int64[] contactIds)
+        public ESIResponse<object> DeleteCharacterContacts(string authToken, int characterId, Int64[] contactIds)
         {
             var request = RestRequestHelper.CreateAuthorizedDeleteRequest($"characters/{characterId}/contacts/", authToken);
             request.AddParameter("contact_ids", contactIds, ParameterType.QueryString);
@@ -52,7 +52,7 @@ namespace EveESIClient.Contacts
         /// <param name="characterId"></param>
         /// <param name="page">Which page of results to return (Default is 0)</param>
         /// <returns></returns>
-        public ESIResponse<List<GetCharacterContactsResponse>> GetCharacterContacts(string authToken, Int64 characterId, int page = 0)
+        public ESIResponse<List<GetCharacterContactsResponse>> GetCharacterContacts(string authToken, int characterId, int page = 0)
         {
             var request = RestRequestHelper.CreateAuthorizedGetRequest($"characters/{characterId}/contacts/", authToken);
 
@@ -69,7 +69,7 @@ namespace EveESIClient.Contacts
         /// <param name="labelId">Add a custom label to the new contact</param>
         /// <param name="watched">Whether the new contact should be watched, note this is only effective on characters</param>
         /// <returns>A list of contact ids that successfully created</returns>
-        public ESIResponse<List<Int64>> AddContacts(string authToken, Int64 characterId, Int64[] contactIds, float standing, int? labelId, bool? watched)
+        public ESIResponse<List<Int64>> AddContacts(string authToken, int characterId, Int64[] contactIds, float standing, int? labelId, bool? watched)
         {
             var request = RestRequestHelper.CreateAuthorizedPostRequest($"characters/{characterId}/contacts/", authToken);
             request.AddParameter("standing", standing, ParameterType.QueryString);
@@ -90,7 +90,7 @@ namespace EveESIClient.Contacts
         /// <param name="labelId">Add a custom label to the new contact</param>
         /// <param name="watched">Whether the new contact should be watched, note this is only effective on characters</param>
         /// <returns>IsSuccessful will be true on successfull call</returns>
-        public ESIResponse<object> UpdateContacts(string authToken, Int64 characterId, Int64[] contactIds, float standing, int? labelId, bool? watched)
+        public ESIResponse<object> UpdateContacts(string authToken, int characterId, Int64[] contactIds, float standing, int? labelId, bool? watched)
         {
             var request = RestRequestHelper.CreateAuthorizedPutRequest($"characters/{characterId}/contacts/", authToken);
             request.AddParameter("standing", standing, ParameterType.QueryString);
@@ -107,7 +107,7 @@ namespace EveESIClient.Contacts
         /// <param name="authToken">Access token to use</param>
         /// <param name="characterId">An EVE character ID</param>
         /// <returns>A list of contact labels</returns>
-        public ESIResponse<List<GetContactLabelsResponse>> GetContactLabels(string authToken, Int64 characterId)
+        public ESIResponse<List<GetContactLabelsResponse>> GetContactLabels(string authToken, int characterId)
         {
             var request = RestRequestHelper.CreateAuthorizedGetRequest($"characters/{characterId}/contacts/labels/", authToken);
 
@@ -121,7 +121,7 @@ namespace EveESIClient.Contacts
         /// <param name="corporationId">An EVE corporation ID</param>
         /// <param name="page">Which page of results to return (Default is 0)</param>
         /// <returns>A list of contacts</returns>
-        public ESIResponse<List<GetCorporationContactsResponse>> GetCorporationContacts(string authToken, Int64 corporationId, int page = 0)
+        public ESIResponse<List<GetCorporationContactsResponse>> GetCorporationContacts(string authToken, int corporationId, int page = 0)
         {
             var request = RestRequestHelper.CreateAuthorizedGetRequest($"corporations/{corporationId}/contacts/", authToken);
