+++ EveESIClient/Fittings/FittingsClient.cs
@@ -21,7 +21,7 @@ namespace EveESIClient.Fittings
         /// <param name="authtoken">Access token to use</param>
         /// <param name="characterId">An EVE character ID</param>
         /// <returns>A list of fittings</returns>
-        public ESIResponse<List<GetCharacterFittingsResponse>> GetCharacterFittings(string authtoken, Int64 characterId)
+        public ESIResponse<List<GetCharacterFittingsResponse>> GetCharacterFittings(string authtoken, int characterId)
         {
             var request = RestRequestHelper.CreateAuthorizedGetRequest($"characters/{characterId}/fittings/", authtoken);
 
@@ -35,7 +35,7 @@ namespace EveESIClient.Fittings
         /// <param name="characterId">An EVE character ID</param>
         /// <param name="input">Details about the new fitting</param>
         /// <returns>Newly created fitting id</returns>
-        public ESIResponse<CreateFittingResponse> CreateFitting(string authToken, Int64 characterId, CreateFittingInput input)
+        public ESIResponse<CreateFittingResponse> CreateFitting(string authToken, int characterId, CreateFittingInput input)
         {
             var request = RestRequestHelper.CreateAuthorizedPostRequest($"characters/{characterId}/fittings/", authToken);
             request.AddBody(input);
@@ -50,7 +50,7 @@ namespace EveESIClient.Fittings
         /// <param name="characterId">An EVE character ID</param>
         /// <param name="fittingId">ID for a fitting of this character</param>
         /// <returns>IsSuccessful will be true on a successful call</returns>
-        public ESIResponse<object> DeleteFitting(string authToken, Int64 characterId, int fittingId)
+        public ESIResponse<object> DeleteFitting(string authToken, int characterId, int fittingId)
         {
             var request = RestRequestHelper.CreateAuthorizedDeleteRequest($"characters/{characterId}/fittings/{fittingId}", authToken);
